query 최적화

+++ src/main/java/ml/psj2867/fancake/service/video/model/BuyStockErrorDto.java
@@ -10,6 +10,6 @@ import ml.psj2867.fancake.service.api.model.ErrorDto;
 @SuperBuilder
 public class BuyStockErrorDto extends ErrorDto{
 
-    private int reaminSize;
+    private long reaminSize;
     
 }
